task 12/RESTURENT CHATBOT  LAB TASKK 12/requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
pandas
